Refactor the sandbox
- improve encapsulation with more static functions
- add lsb prefix to any exposed defines/functions/types
- break out lua modules so they can be used outside the sandbox
  - add an interface to allow modules to interact with sandbox
    serialization and output
