--- src/cephes.h
@@ -1,8 +1,0 @@
-/*
-* Cephes Math Library Release 2.2: June, 1992
-* Copyright 1984, 1987, 1988, 1992 by Stephen L. Moshier
-* Direct inquiries to 30 Frost Street, Cambridge, MA 02140
-*/
-
-/* Normal Distribution Function */
-double ndtr(double a);
