+++ src/test/test_lua_sandbox.c
@@ -6,15 +6,16 @@
 
 /** @brief Lua sandbox unit tests @file */
 
-#include "lsb.h"
-
 #include <errno.h>
-#include <lua.h>
 #include <lauxlib.h>
+#include <lua.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
+
+#include "lsb.h"
+#include "lsb_output.h"
 
 #ifdef _WIN32
 #define snprintf _snprintf
@@ -200,10 +201,6 @@ int write_message(lua_State* lua)
   }
   lua_sandbox* lsb = (lua_sandbox*)luserdata;
 
-  if (lua_gettop(lua) != 1 || lua_type(lua, 1) != LUA_TTABLE) {
-    luaL_error(lua, "write_message() takes a table argument");
-  }
-
   if (lsb_output_protobuf(lsb, 1, 0) != 0) {
     luaL_error(lua, "write_message() could not encode protobuf - %s",
                lsb_get_error(lsb));
@@ -269,7 +266,7 @@ static char* test_init_error()
 
 static char* test_destroy_error()
 {
-  const char* expected = "preserve_global_data could not open: "
+  const char* expected = "lsb_preserve_global_data could not open: "
     "invaliddir/simple.preserve";
   e = lsb_destroy(NULL, NULL);
   mu_assert(!e, "lsb_destroy() received: %s", e);
@@ -491,6 +488,7 @@ static char* test_output_errors()
     , "process() lua/output_errors.lua:27: output_limit exceeded"
     , "process() lua/output_errors.lua:30: write_message() could not encode protobuf - unsupported array type: table"
     , "process() lua/output_errors.lua:36: strbuf max_size exceeded"
+    , "process() lua/output_errors.lua:38: write_message() could not encode protobuf - takes a single table argument"
     , NULL
   };
 
@@ -524,44 +522,7 @@ static char* test_cbuf_errors()
 {
   const char* tests[] =
   {
-    "process() lua/circular_buffer_errors.lua:9: bad argument #0 to 'new' (incorrect number of arguments)"
-    , "process() lua/circular_buffer_errors.lua:11: bad argument #1 to 'new' (number expected, got nil)"
-    , "process() lua/circular_buffer_errors.lua:13: bad argument #1 to 'new' (rows must be > 1)"
-    , "process() lua/circular_buffer_errors.lua:15: bad argument #2 to 'new' (number expected, got nil)"
-    , "process() lua/circular_buffer_errors.lua:17: bad argument #2 to 'new' (columns must be > 0)"
-    , "process() lua/circular_buffer_errors.lua:19: bad argument #3 to 'new' (number expected, got nil)"
-    , "process() lua/circular_buffer_errors.lua:21: bad argument #3 to 'new' (seconds_per_row is out of range)"
-    , "process() lua/circular_buffer_errors.lua:23: removed test"
-    , "process() not enough memory"
-    , "process() lua/circular_buffer_errors.lua:28: bad argument #2 to 'set' (column out of range)"
-    , "process() lua/circular_buffer_errors.lua:31: bad argument #2 to 'set' (column out of range)"
-    , "process() lua/circular_buffer_errors.lua:34: bad argument #2 to 'set' (number expected, got nil)"
-    , "process() lua/circular_buffer_errors.lua:37: bad argument #1 to 'set' (number expected, got nil)"
-    , "process() lua/circular_buffer_errors.lua:41: bad argument #1 to 'get' (lsb.circular_buffer expected, got number)"
-    , "process() lua/circular_buffer_errors.lua:44: bad argument #3 to 'set' (number expected, got nil)"
-    , "process() lua/circular_buffer_errors.lua:47: bad argument #-1 to 'set' (incorrect number of arguments)"
-    , "process() lua/circular_buffer_errors.lua:50: bad argument #-1 to 'add' (incorrect number of arguments)"
-    , "process() lua/circular_buffer_errors.lua:53: bad argument #-1 to 'get' (incorrect number of arguments)"
-    , "process() lua/circular_buffer_errors.lua:56: bad argument #-1 to 'compute' (incorrect number of arguments)"
-    , "process() lua/circular_buffer_errors.lua:59: bad argument #1 to 'compute' (invalid option 'func')"
-    , "process() lua/circular_buffer_errors.lua:62: bad argument #2 to 'compute' (column out of range)"
-    , "process() lua/circular_buffer_errors.lua:65: bad argument #4 to 'compute' (end must be >= start)"
-    , "process() lua/circular_buffer_errors.lua:68: bad argument #1 to 'format' (invalid option 'invalid')"
-    , "process() lua/circular_buffer_errors.lua:71: bad argument #-1 to 'format' (incorrect number of arguments)"
-    , "process() lua/circular_buffer_errors.lua:74: bad argument #-1 to 'format' (incorrect number of arguments)"
-    , "process() lua/circular_buffer_errors.lua:77: fromstring() too few values: 0, expected 2"
-    , "process() lua/circular_buffer_errors.lua:80: fromstring() too few values: 0, expected 2"
-    , "process() lua/circular_buffer_errors.lua:83: fromstring() too many values, more than: 2"
-    , "process() lua/circular_buffer_errors.lua:86: bad argument #-1 to 'mannwhitneyu' (incorrect number of arguments)"
-    , "process() lua/circular_buffer_errors.lua:89: bad argument #1 to 'mannwhitneyu' (number expected, got nil)"
-    , "process() lua/circular_buffer_errors.lua:92: bad argument #1 to 'mannwhitneyu' (column out of range)"
-    , "process() lua/circular_buffer_errors.lua:95: bad argument #3 to 'mannwhitneyu' (ranges must not overlap)"
-    , "process() lua/circular_buffer_errors.lua:98: bad argument #3 to 'mannwhitneyu' (end_1 must be >= start_1)"
-    , "process() lua/circular_buffer_errors.lua:101: bad argument #5 to 'mannwhitneyu' (end_2 must be >= start_2)"
-    , "process() lua/circular_buffer_errors.lua:104: bad argument #-1 to 'mannwhitneyu' (too many arguments)"
-    , "process() lua/circular_buffer_errors.lua:107: bad argument #6 to 'mannwhitneyu' (use_continuity must be a boolean)"
-    , "process() lua/circular_buffer_errors.lua:110: bad argument #-1 to 'get_header' (incorrect number of arguments)"
-    , "process() lua/circular_buffer_errors.lua:113: bad argument #1 to 'get_header' (column out of range)"
+    "process() not enough memory"
     , NULL
   };
 
@@ -637,12 +598,6 @@ static char* test_cbuf()
   mu_assert(result == 0, "report() received: %d", result);
   mu_assert(strcmp(outputs[3], written_data) == 0, "received: %s",
             written_data);
-
-  for (int i = 1; i < 18; ++i) {
-    result = report(sb, i);
-    mu_assert(result == 0, "report() test: %d received: %d error: %s", i,
-              result, lsb_get_error(sb));
-  }
 
   e = lsb_destroy(sb, "circular_buffer.preserve");
   mu_assert(!e, "lsb_destroy() received: %s", e);
@@ -972,7 +927,7 @@ static char* test_serialize_failure()
 static char* test_serialize_noglobal()
 {
   const char* output_file = "serialize_noglobal.preserve";
-  const char* expected = "preserve_global_data cannot access the global table";
+  const char* expected = "lsb_preserve_global_data cannot access the global table";
 
   lua_sandbox* sb = lsb_create(NULL,
                                "lua/serialize_noglobal.lua", "../../modules",
@@ -988,50 +943,6 @@ static char* test_serialize_noglobal()
   free(e);
   e = NULL;
   mu_assert(file_exists(output_file) == 0, "output file was not cleaned up");
-
-  return NULL;
-}
-
-
-static char* test_bloom_filter_errors()
-{
-  const char* tests[] =
-  {
-    "process() lua/bloom_filter_errors.lua:9: bad argument #0 to 'new' (incorrect number of arguments)"
-    , "process() lua/bloom_filter_errors.lua:11: bad argument #1 to 'new' (number expected, got nil)"
-    , "process() lua/bloom_filter_errors.lua:13: bad argument #1 to 'new' (items must be > 1)"
-    , "process() lua/bloom_filter_errors.lua:15: bad argument #2 to 'new' (number expected, got nil)"
-    , "process() lua/bloom_filter_errors.lua:17: bad argument #2 to 'new' (probability must be between 0 and 1)"
-    , "process() lua/bloom_filter_errors.lua:19: bad argument #2 to 'new' (probability must be between 0 and 1)"
-    , "process() lua/bloom_filter_errors.lua:22: bad argument #-1 to 'add' (incorrect number of arguments)"
-    , "process() lua/bloom_filter_errors.lua:25: bad argument #1 to 'add' (must be a string or number)"
-    , "process() lua/bloom_filter_errors.lua:28: bad argument #-1 to 'query' (incorrect number of arguments)"
-    , "process() lua/bloom_filter_errors.lua:31: bad argument #1 to 'query' (must be a string or number)"
-    , "process() lua/bloom_filter_errors.lua:34: bad argument #-1 to 'clear' (incorrect number of arguments)"
-    , "process() lua/bloom_filter_errors.lua:37: bad argument #1 to 'fromstring' (string expected, got table)"
-    , "process() lua/bloom_filter_errors.lua:40: fromstring() bytes found: 23, expected 24"
-    , NULL
-  };
-
-  for (int i = 0; tests[i]; ++i) {
-    lua_sandbox* sb = lsb_create(NULL, "lua/bloom_filter_errors.lua",
-                                 "../../modules", 32767, 1000, 128);
-    mu_assert(sb, "lsb_create() received: NULL");
-
-    int result = lsb_init(sb, NULL);
-    mu_assert(result == 0, "lsb_init() received: %d %s", result,
-              lsb_get_error(sb));
-
-    result = process(sb, i);
-    mu_assert(result == 1, "test: %d received: %d", i, result);
-
-    const char* le = lsb_get_error(sb);
-    mu_assert(le, "test: %d received NULL", i);
-    mu_assert(strcmp(tests[i], le) == 0, "test: %d received: %s", i, le);
-
-    e = lsb_destroy(sb, NULL);
-    mu_assert(!e, "lsb_destroy() received: %s", e);
-  }
 
   return NULL;
 }
@@ -1106,45 +1017,6 @@ static char* test_bloom_filter()
 
   e = lsb_destroy(sb, NULL);
   mu_assert(!e, "lsb_destroy() received: %s", e);
-
-  return NULL;
-}
-
-
-static char* test_hyperloglog_errors()
-{
-  const char* tests[] =
-  {
-    "process() lua/hyperloglog_errors.lua:9: bad argument #0 to 'new' (incorrect number of arguments)"
-    , "process() lua/hyperloglog_errors.lua:12: bad argument #1 to 'add' (must be a string or number)"
-    , "process() lua/hyperloglog_errors.lua:15: bad argument #-1 to 'add' (incorrect number of arguments)"
-    , "process() lua/hyperloglog_errors.lua:18: bad argument #-1 to 'count' (incorrect number of arguments)"
-    , "process() lua/hyperloglog_errors.lua:21: bad argument #-1 to 'clear' (incorrect number of arguments)"
-    , "process() lua/hyperloglog_errors.lua:24: bad argument #1 to 'fromstring' (string expected, got table)"
-    , "process() lua/hyperloglog_errors.lua:27: fromstring() bytes found: 23, expected 12304"
-    , NULL
-  };
-
-  for (int i = 0; tests[i]; ++i) {
-    lua_sandbox* sb = lsb_create(NULL,
-                                 "lua/hyperloglog_errors.lua", "../../modules",
-                                 32767, 1000, 128);
-    mu_assert(sb, "lsb_create() received: NULL");
-
-    int result = lsb_init(sb, NULL);
-    mu_assert(result == 0, "lsb_init() received: %d %s", result,
-              lsb_get_error(sb));
-
-    result = process(sb, i);
-    mu_assert(result == 1, "test: %d received: %d", i, result);
-
-    const char* le = lsb_get_error(sb);
-    mu_assert(le, "test: %d received NULL", i);
-    mu_assert(strcmp(tests[i], le) == 0, "test: %d received: %s", i, le);
-
-    e = lsb_destroy(sb, NULL);
-    mu_assert(!e, "lsb_destroy() received: %s", e);
-  }
 
   return NULL;
 }
@@ -1530,9 +1402,7 @@ static char* all_tests()
   mu_run_test(test_restore);
   mu_run_test(test_serialize_failure);
   mu_run_test(test_serialize_noglobal);
-  mu_run_test(test_bloom_filter_errors);
   mu_run_test(test_bloom_filter);
-  mu_run_test(test_hyperloglog_errors);
   mu_run_test(test_hyperloglog);
   mu_run_test(test_struct);
